AI-Sketch-to-realistic-maker/requirements.txt
gradio
pillow
requests
transformers
torch
diffusers
accelerate
sentence_transformers